Fix FK type mismatch in marcacoes_em_falta migration
marcacao_entrada_id era INT UNSIGNED mas marcacoes.id e BIGINT UNSIGNED.
MySQL rejeitava a FK constraint com ERROR 3780 (HY000).
Migration aplicada manualmente em tenant_004_ftl apos correccao.

+++ database/migrations/tenant/003_marcacoes_em_falta.sql
@@ -5,7 +5,7 @@ CREATE TABLE marcacoes_em_falta (
     id INT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     funcionario_id INT UNSIGNED NOT NULL,
     data DATE NOT NULL,
-    marcacao_entrada_id INT UNSIGNED NULL,
+    marcacao_entrada_id BIGINT UNSIGNED NULL,
     estado ENUM(
         'pendente',
         'justificada_trabalho',
